ADD: support recover checkpoint

+++ src/main/java/org/example/nem/data/ErrorRecord.java
@@ -2,5 +2,5 @@ package org.example.nem.data;
 
 import org.example.nem.constant.ErrorType;
 
-public record ErrorRecord(String inputFileName, int lineNumber, ErrorType errorType) {
+public record ErrorRecord(String inputFileName, long lineNumber, ErrorType errorType) {
 }
